requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.3
langchain_community==0.3.2
langchain_core==0.3.11
langchain_huggingface==0.1.0
python-dotenv==1.0.1
torch==2.3.0
torch==2.4.1+cu121
transformers==4.45.0.dev0
